+++ components/site-header.tsx
@@ -12,9 +12,6 @@ import {
   BreadcrumbSeparator,
 } from '@/components/ui/breadcrumb';
 import { usePathname, useRouter } from 'next/navigation';
-import { Button } from './ui/button';
-import { GPClient } from '@/axios-instance';
-import { toast } from 'sonner';
 
 export function SiteHeader() {
   const router = useRouter();
@@ -30,22 +27,15 @@ export function SiteHeader() {
       return { href, label };
     });
 
-  async function logout() {
-    await GPClient.post('/api/auth/logout');
-    localStorage.removeItem('user_info');
-    toast.success('Cerraste sesión');
-    router.refresh();
-  }
-
   return (
-    <header className="flex justify-between bg-background sticky top-0 flex h-16 shrink-0 items-center gap-2 border-b px-4 z-1">
+    <header className="bg-background sticky top-0 flex h-16 shrink-0 items-center gap-2 border-b px-4 z-1">
       <div className="flex align-center items-center">
         <SidebarTrigger className="-ml-1" />
         <Separator orientation="vertical" className="mr-2 h-4" />
         <Breadcrumb>
           <BreadcrumbList>
             <BreadcrumbItem className="hidden md:block">
-              <BreadcrumbLink href="/">MyGP</BreadcrumbLink>
+              <BreadcrumbLink href="/transbel/dashboard">MyGP</BreadcrumbLink>
             </BreadcrumbItem>
             {breadcrumbs.map((crumb, index) => (
               <React.Fragment key={crumb.href}>
@@ -61,11 +51,6 @@ export function SiteHeader() {
             ))}
           </BreadcrumbList>
         </Breadcrumb>
-      </div>
-      <div>
-        <Button className="bg-red-500 hover:bg-red-600 cursor-pointer" onClick={logout}>
-          Cerrar Sesión
-        </Button>
       </div>
     </header>
   );
